添加PermissionService方法

+++ src/main/java/com/cnsunet/kjw/controller/sysmanage/UserController.java
@@ -8,6 +8,7 @@ import com.cnsunet.kjw.service.sysmanage.impl.UserService;
 import com.cnsunet.kjw.utils.StatusDefine;
 import com.cnsunet.kjw.utils.StatusDefineMessage;
 import com.cnsunet.kjw.utils.json.JsonResponseData;
+import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import io.swagger.annotations.ApiOperation;
 import io.swagger.annotations.ApiParam;
 import org.slf4j.Logger;
@@ -51,13 +52,28 @@ public class UserController /*extends AbstractAuthenticationProcessingFilter*/ {
     @Autowired
     private CustomAuthenticationDetailsSource customAuthenticationDetailsSource;
 
+
     @RequestMapping(value = "/api/user",method = RequestMethod.GET)
     @ApiOperation(value = "",response = String.class,httpMethod = "GET",notes = "")
     public String getUserById(
             @ApiParam(value = "用户Id",required = true) @RequestParam(value = "id",required = true) String id
     ){
-        UserModel userModel=userService.getUserById(Integer.valueOf(id));
-        return new JsonResponseData(true,"UserController-getUserById",200,"success",userModel).toString();
+
+        try {
+            UserModel result=userService.getUserById(Integer.valueOf(id));
+            if (result==null) {
+                return new JsonResponseData(false, StatusDefineMessage.GetMessage(StatusDefine.U_INEXISTENCE), StatusDefine.U_INEXISTENCE, "修改用户失败", null).toString();
+            }
+            return new JsonResponseData(true, StatusDefineMessage.GetMessage(StatusDefine.SUCCESS), StatusDefine.SUCCESS, "获取用户信息成功", result).toString();
+        }catch (DBErrorException e){
+            //抛出异常返回异常信息
+            logger.error("controller:UserController. function:getUserById..msg:GET  DBErrorException. error:"+e.getMessage());
+            return new JsonResponseData(false, StatusDefineMessage.GetMessage(StatusDefine.DB_ERROR), StatusDefine.DB_ERROR, "", null).toString();
+        }catch (Exception e){
+            //抛出异常返回异常信息
+            logger.error("controller:UserController. function:getUserById..msg:GET  Exception. error:"+e.getMessage());
+            return new JsonResponseData(false, StatusDefineMessage.GetMessage(StatusDefine.SYS_ERROR), StatusDefine.SYS_ERROR, "", null).toString();
+        }
     }
     /**
      * @Author  huangjie
@@ -70,7 +86,7 @@ public class UserController /*extends AbstractAuthenticationProcessingFilter*/ {
      */
     @RequestMapping(value = "/api/login",method = RequestMethod.GET)
     @ApiOperation(value = "登陆",httpMethod = "GET",response = String.class,notes = "登陆")
-    public String login(
+    public String userLogin(
             @ApiParam(value = "用户名",required =true ) @RequestParam(value = "username",required = true) String username,
             @ApiParam(value = "密码",required =true ) @RequestParam(value = "password",required = true) String password,
             @ApiParam(value = "token1",required =true ) @RequestParam(value = "token",required = true) String token,
@@ -101,9 +117,9 @@ public class UserController /*extends AbstractAuthenticationProcessingFilter*/ {
      */
     @RequestMapping(value = "/api/logout",method = RequestMethod.PUT)
     @ApiOperation(value = "退出登录",httpMethod = "PUT",response = String.class,notes = "登出")
-    public String logout(HttpServletRequest request){
+    public String userLogout(HttpServletRequest request){
         if (request.getSession().getAttribute("SPRING_SECURITY_CONTEXT")==null) {
-            return new JsonResponseData(true, StatusDefineMessage.GetMessage(StatusDefine.SUCCESS), StatusDefine.SUCCESS, "用户未登录", null).toString();
+            return new JsonResponseData(false, StatusDefineMessage.GetMessage(StatusDefine.U_UNLOAD), StatusDefine.U_UNLOAD, "用户未登录", null).toString();
         }
         SecurityContextHolder.getContext().setAuthentication(null);
         HttpSession session = request.getSession();
@@ -124,12 +140,18 @@ public class UserController /*extends AbstractAuthenticationProcessingFilter*/ {
      @ApiOperation(value ="添加用户",response = String.class,httpMethod = "POST",notes="添加用户-供管理员添加web登陆用户使用")
      public String addUser(
              @ApiParam(value = "用户名",required = true ) @RequestParam(value = "userName" ,required = true) String userName,
-             @ApiParam(value = "用户密码",required = true) @RequestParam(value = "password",required = true) String password
+             @ApiParam(value = "用户密码",required = true) @RequestParam(value = "password",required = true) String password,
+             @ApiParam(value = "创建者Id",required = true) @RequestParam(value = "createId",required = true) String createId
      ){
          UserModel userModel=new UserModel();
          userModel.setUserName(userName);
          userModel.setPassword(password);
+         userModel.setCreateId(Integer.valueOf(createId));
          try {
+             UserModel old=userService.getUserByName(userName);
+             if(!(old==null)){
+                 return new JsonResponseData(false, StatusDefineMessage.GetMessage(StatusDefine.U_EXIST_USER), StatusDefine.U_EXIST_USER, "添加用户失败", null).toString();
+             }
              Integer result=userService.addUser(userModel);
              return new JsonResponseData(true, StatusDefineMessage.GetMessage(StatusDefine.SUCCESS), StatusDefine.SUCCESS, "添加用户成功", result).toString();
          }catch (DBErrorException e){
@@ -155,7 +177,7 @@ public class UserController /*extends AbstractAuthenticationProcessingFilter*/ {
     @PreAuthorize("hasUserAddPms()")
     @RequestMapping(value = "/api/updateforadmin",method = RequestMethod.PUT)
     @ApiOperation(value ="修改用户FORADMIN",response = String.class,httpMethod = "PUT",notes="修改用户-供管理员修改web登陆用户使用")
-    public String updateUserFroAdmin(
+    public String updateUserForAdmin(
             @ApiParam(value = "用户id",required = true ) @RequestParam(value = "userId" ,required = true) String userId,
             @ApiParam(value = "用户名",required = true ) @RequestParam(value = "userName" ,required = true) String userName,
             @ApiParam(value = "用户密码",required = true) @RequestParam(value = "password",required = true) String password
@@ -165,6 +187,10 @@ public class UserController /*extends AbstractAuthenticationProcessingFilter*/ {
         userModel.setUserName(userName);
         userModel.setPassword(password);
         try {
+            UserModel old=userService.getUserById(Integer.valueOf(userId));
+            if (old==null) {
+                return new JsonResponseData(false, StatusDefineMessage.GetMessage(StatusDefine.U_INEXISTENCE), StatusDefine.U_INEXISTENCE, "修改用户失败", null).toString();
+            }
             Integer result=userService.updateUser(userModel);
             return new JsonResponseData(true, StatusDefineMessage.GetMessage(StatusDefine.SUCCESS), StatusDefine.SUCCESS, "修改用户成功", result).toString();
         }catch (DBErrorException e){
@@ -188,7 +214,7 @@ public class UserController /*extends AbstractAuthenticationProcessingFilter*/ {
      */
     @RequestMapping(value = "/api/updateforuser",method = RequestMethod.PUT)
     @ApiOperation(value ="修改用户FORUSER",response = String.class,httpMethod = "PUT",notes="修改用户-供web登陆用户使用不需要权限")
-    public String updateUserFroUser(
+    public String updateUserForUser(
             @ApiParam(value = "用户id",required = true ) @RequestParam(value = "userId" ,required = true) String userId,
             @ApiParam(value = "用户名",required = true ) @RequestParam(value = "userName" ,required = true) String userName,
             @ApiParam(value = "用户密码",required = true) @RequestParam(value = "password",required = true) String password
@@ -198,6 +224,10 @@ public class UserController /*extends AbstractAuthenticationProcessingFilter*/ {
         userModel.setUserName(userName);
         userModel.setPassword(password);
         try {
+            UserModel old=userService.getUserById(Integer.valueOf(userId));
+            if (old==null) {
+                return new JsonResponseData(false, StatusDefineMessage.GetMessage(StatusDefine.U_INEXISTENCE), StatusDefine.U_INEXISTENCE, "修改用户失败", null).toString();
+            }
             Integer result=userService.updateUser(userModel);
             return new JsonResponseData(true, StatusDefineMessage.GetMessage(StatusDefine.SUCCESS), StatusDefine.SUCCESS, "修改用户成功", result).toString();
         }catch (DBErrorException e){
@@ -220,21 +250,25 @@ public class UserController /*extends AbstractAuthenticationProcessingFilter*/ {
      * @Modyfied by
      */
     @PreAuthorize("hasUserDeletePms()")
-    @RequestMapping(value = "/api/deleteuser",method = RequestMethod.PUT)
-    @ApiOperation(value ="删除用户",response = String.class,httpMethod = "PUT",notes="删除用户-供管理员删除用户使用需要权限")
+    @RequestMapping(value = "/api/deleteuser",method = RequestMethod.DELETE)
+    @ApiOperation(value ="删除用户",response = String.class,httpMethod = "DELETE",notes="删除用户-供管理员删除用户使用需要权限")
     public String deleteUser(
             @ApiParam(value = "用户名",required = true ) @RequestParam(value = "userName" ,required = true) String userName
     ){
         try {
+            UserModel user=userService.getUserByName(userName);
+            if (user==null) {
+                return new JsonResponseData(false, StatusDefineMessage.GetMessage(StatusDefine.U_INEXISTENCE), StatusDefine.U_INEXISTENCE, "删除用户失败", null).toString();
+            }
             Integer result=userService.deleteUserByName(userName);
             return new JsonResponseData(true, StatusDefineMessage.GetMessage(StatusDefine.SUCCESS), StatusDefine.SUCCESS, "删除用户成功", result).toString();
         }catch (DBErrorException e){
             //抛出异常返回异常信息
-            logger.error("controller:UserController. function:deleteUser..msg:PUT  DBErrorException. error:"+e.getMessage());
+            logger.error("controller:UserController. function:deleteUser..msg:DELETE  DBErrorException. error:"+e.getMessage());
             return new JsonResponseData(false, StatusDefineMessage.GetMessage(StatusDefine.DB_ERROR), StatusDefine.DB_ERROR, "", null).toString();
         }catch (Exception e){
             //抛出异常返回异常信息
-            logger.error("controller:UserController. function:deleteUser..msg:PUT  Exception. error:"+e.getMessage());
+            logger.error("controller:UserController. function:deleteUser..msg:DELETE  Exception. error:"+e.getMessage());
             return new JsonResponseData(false, StatusDefineMessage.GetMessage(StatusDefine.SYS_ERROR), StatusDefine.SYS_ERROR, "", null).toString();
         }
     }
@@ -248,6 +282,7 @@ public class UserController /*extends AbstractAuthenticationProcessingFilter*/ {
      * 方法异常 DBErrorException  Exception
      * @Modyfied by
      */
+    @PreAuthorize("hasSadmin()")
     @RequestMapping(value = "/api/userAll",method = RequestMethod.GET)
     @ApiOperation(value ="获取所有用户信息",response = String.class,httpMethod = "GET",notes="获取所有用户信息-供管理员使用")
     public String getAllUser(){
@@ -273,12 +308,17 @@ public class UserController /*extends AbstractAuthenticationProcessingFilter*/ {
      * 方法异常 DBErrorException  Exception
      * @Modyfied by
      */
+    @PreAuthorize("hasRoleSelectPms()")
     @RequestMapping(value = "/api/userAllFroCreator",method = RequestMethod.GET)
     @ApiOperation(value ="根据用户id获取该用户所创建的所有用户的信息",response = String.class,httpMethod = "GET",notes="根据用户id获取该用户所创建的所有用户的信息-供web登陆者使用需权限")
     public String getAllUserForCreator(
             @ApiParam(value = "创建者ID",required = true ) @RequestParam(value = "createId" ,required = true) String createId
     ){
         try {
+            UserModel user=userService.getUserById(Integer.valueOf(createId));
+            if (user==null) {
+                return new JsonResponseData(false, StatusDefineMessage.GetMessage(StatusDefine.U_INEXISTENCE), StatusDefine.U_INEXISTENCE, "删除用户失败", null).toString();
+            }
             List<UserModel> result=userService.getAllUserByCreatUser(Integer.valueOf(createId));
             return new JsonResponseData(true, StatusDefineMessage.GetMessage(StatusDefine.SUCCESS), StatusDefine.SUCCESS, "根据用户id获取该用户所创建的所有用户成功", result).toString();
         }catch (DBErrorException e){
